requirements.txt
fastapi==0.118.0
streamlit==1.39.0
uvicorn==0.37.0
langchain==0.3.27
langchain-chroma==0.2.6
langchain-community==0.3.31
langchain-core==0.3.79
langchain-google-genai==2.1.12
langchain-text-splitters==0.3.11
langgraph==0.6.10
chromadb==1.1.1
sentence-transformers==5.1.1
torch==2.9.0
transformers==4.57.1
scikit-learn==1.7.2
requests==2.32.5
python-multipart==0.0.6
python-dotenv==1.1.1
google-ai-generativelanguage==0.7.0